Abstracted function solution for atoi,
This function is abstracted by replacing str[indx] with *str and using indx++ counter has been shortened to *(str++). in addition, some variable names are changed to shorten the lines used because or the Norm. This is a single function that i didnt use because it may be tricky for me to read in the future.